+++ docs/changelog.md
@@ -15,6 +15,12 @@ Dates are specified in the format `DD-MM-YYYY`.
   - RedSun -> Redsun
   - SunFlare -> Sunflare
 
+### Added
+
+- Added documentation
+  - Notions of O.O.P.
+  - Minimal model doc 
+
 ## [v0.3.2] - 29-01-2025
 
 ### Added